LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowframefrom.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <body>
    <script src="../resources/securitypolicy-tests-base.js"></script>
    <script>
      if (document.SecurityPolicy.allowsFrameFrom('http://example.com/'))
          log('PASS frame is allowed when no policy exists.');
      else
          log('FAIL frame is not allowed when no policy exists.');

      injectPolicy("frame-src http://notexample.com;");

      if (!document.SecurityPolicy.allowsFrameFrom('http://example.com/'))
          log('PASS frame is not allowed when policy exists.');
      else
          log('FAIL frame is allowed when policy exists.');
    </script>
  </body>
</html>
