safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowframefrom-expected.txt

PASS frame is allowed when no policy exists.
PASS frame is not allowed when policy exists.
